requirements.txt
PyPDF2
bs4
contractions
google-generativeai
langchain
langchain-community
langchain-google-genai
nltk
python-dotenv
regex
unidecode
sentence-transformers==2.2.2
